+++ RELEASES.md
@@ -23,3 +23,11 @@ Based on:
 - Speakeasy CLI 1.300.0 (2.338.14) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v0.0.5] ./packages/client-reconciliation
+
+## 2024-06-07 14:31:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.300.1 (2.339.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.6] ./packages/client-ledger